risk_analyzer/requirements.txt
fastapi
uvicorn[standard]
httpx
numpy
pandas
scipy
python-dotenv
